Work a good bit on markup, do some silly cosmetic stuff
