+++ assets/src/Ellie/Ui/CodeEditor.js
@@ -63,7 +63,7 @@ const load = () => {
     import(/* webpackChunkName: "codemirror-base" */ "codemirror/lib/codemirror"),
     import(/* webpackChunkName: "codemirror-base", webpackMode: "eager" */ "codemirror/lib/codemirror.css"),
     import(/* webpackChunkName: "codemirror-base" */ "codemirror/mode/elm/elm"),
-    import(/* webpackChunkName: "codemirror-base" */ "./Markup"),
+    import(/* webpackChunkName: "codemirror-base" */ "./BrillinkMode"),
     import(/* webpackChunkName: "codemirror-base" */ "codemirror/mode/gfm/gfm"),
     import(/* webpackChunkName: "codemirror-base" */ "codemirror/addon/lint/lint"),
     import(/* webpackChunkName: "codemirror-base" */ "codemirror/addon/selection/active-line"),
