+++ assets/src/Pages/Embed/Views/App.elm
@@ -192,7 +192,7 @@ viewContent state =
             Panel.Html ->
                 viewOverlayed <|
                     CodeEditor.view
-                        [ CodeEditor.mode "gfm"
+                        [ CodeEditor.mode "brillink"
                         , CodeEditor.value state.revision.data.htmlCode
                         , CodeEditor.readOnly
                         ]
