+++ lib/ellie_web/templates/page/editor.html.eex
@@ -3,6 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="public_path" content="<%= static_url(@conn, "/assets/") %>" />
+    <link href="https://fonts.googleapis.com/css?family=Source+Code+Pro&display=swap" rel="stylesheet">
     <link rel="icon" href="<%= static_url(@conn, "/assets/images/favicon.ico") %>" />
     <link rel="alternate" type="application/json+oembed" href="<%= oembed_url(@conn) %>" />
     <%= render("polyfills.html") %>
