+++ lib/ellie_web/templates/page/embed.html.eex
@@ -3,6 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="public_path" content="<%= static_url(@conn, "/assets/") %>" />
+    <link href="https://fonts.googleapis.com/css?family=Source+Code+Pro&display=swap" rel="stylesheet">
     <link rel="icon" href="<%= static_url(@conn, "/assets/images/favicon.ico") %>" />
     <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1" />
     <%= render("polyfills.html") %>
